[ja] Fix translation on improvements in the Git UI
The Japanese translation implies that all systems have had improvements
in their UI over the past few years, but many aspects of Git are still
hard to use. This is not in agreement with the original English:

"In the last few years, the UI has been refined until it’s as clean and
easy to use as any system out there; but often, the stereotype lingers
about the early Git UI that was complex and difficult to learn."

Change the Japanese translation to more closely reflect the English,
i.e.: that the Git UI has improved, but its image as a hard-to-use tool
still lingers.