[ko] update chapter 1
This commit improves the Korean translation of the first two paragraphs
in Chapter 1. It makes the sentences sound more natural and removes the
wrong clause in the first paragragh which says there will be an
explanation in this chapter for how to set up a Git server, which is not
true.